+++ sqlalchemy_utils/functions/__init__.py
@@ -1,13 +1,19 @@
 from .defer_except import defer_except
 from .mock import create_mock_engine, mock_engine
 from .render import render_expression, render_statement
-from .sort_query import sort_query, QuerySorterException
+from .sort_query import (
+    make_order_by_deterministic,
+    sort_query,
+    QuerySorterException
+)
 from .database import (
+    analyze,
     create_database,
     database_exists,
     drop_database,
     escape_like,
     has_index,
+    has_unique_index,
     is_auto_assigned_date_column,
 )
 from .foreign_keys import (
@@ -61,6 +67,7 @@ __all__ = (
     'identity',
     'is_auto_assigned_date_column',
     'is_indexed_foreign_key',
+    'make_order_by_deterministic',
     'mock_engine',
     'naturally_equivalent',
     'non_indexed_foreign_keys',